port_joystick: Ignore synthetic init events. Only call perror if errno is set.
Pressed button states seems to be preserved between runs of the program somehow.
Any buttons pressed when the program quits will get a JS_EVENT_INIT press event
when the program is run again, even if the buttons are not actually down.
In this case non-synthetic release events follow shortly after the sythetic
events.  This was seen with a Saitek P880 and Linux kernel 4.18.19-100.fc27.